Build required examples for selenium test server
