+++ package.json
@@ -76,7 +76,7 @@
     "ci": "karma start karma-cov.conf.js --single-run --browsers PhantomJS",
     "codecov": "cat lcov/*/lcov.info | codecov",
     "examples": "webpack-dev-server --config webpack-examples.config.js --port 8082 --content-base dist/",
-    "start-test": "forever start ./testing/test-runners/server.js",
+    "start-test": "node examples/build.js; forever start ./testing/test-runners/server.js",
     "stop-test": "forever stop ./testing/test-runners/server.js",
     "docs": "jsdoc --pedantic -d dist/apidocs -r src"
   },
